Fix flake8 errors (drop Python2.7 support)
