+++ queries/remix.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from xdfile.utils import get_args, open_output, find_files, log, debug, info, error, get_log, COLUMN_SEPARATOR, EOL
-from xdfile.utils import parse_tsv, progress, parse_pathname
+from xdfile.utils import parse_tsv, progress, parse_pathname, iso8601
 from xdfile import corpus, xdfile, BLOCK_CHAR
 
 
@@ -94,7 +94,7 @@ def mutate(xd, words, chance=1):
 
 def load_clues():
     ret = {} # ["pubid"] = { ["ANSWER"] = { ["simplified clue text"] = set(fullclues) } }
-    for r in parse_tsv(file("clues.tsv").read(), "AnswerClue"):
+    for r in parse_tsv(open("clues.tsv").read(), "AnswerClue"):
         try:
             pubid, dt, answer, clue = r
         except Exception as e:
