+++ scripts/21-clean-metadata.py
@@ -133,7 +133,7 @@ def clean_headers(xd):
                 dt = d.strftime("%Y-%m-%d")
         except Exception as e:
             utils.error(str(e))
-            if args.debug:
+            if utils.get_args().debug:
                 raise
 
     ## try getting Date from copyright
